Move file to separated folder
